check for existence before deleting links. This prevents failure on first run

+++ open-xchange-munin-scripts-jolokia/open-xchange-munin-scripts-jolokia.spec
@@ -50,7 +50,9 @@ if [ -f $TMPFILE ] ; then
   sh < $TMPFILE
   rm -f $TMPFILE
 fi
-find -L /etc/munin/plugins/ox_* -type l -delete
+if [ -f /etc/munin/plugins/ox_* ] ; then
+  find -L /etc/munin/plugins/ox_* -type l -delete
+fi
 /etc/init.d/munin-node restart || :
 exit 0
